+++ packages/webpack-plugin/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @posthog/webpack-plugin
 
+## 1.4.5
+
+### Patch Changes
+
+- Updated dependencies [[`922a1c1`](https://github.com/PostHog/posthog-js/commit/922a1c1838a5ed2ad37f59dade5fc3cc81bb4246)]:
+    - @posthog/core@1.26.0
+
 ## 1.4.4
 
 ### Patch Changes